changed quadratic file format
so that it aligns with the smps file format. The updated version has a separate set of cor. tim. sto. files for quadratic constraints, which extends the smps files. Integrating the quadratic file format into the smps file seems to require some amount of work as it may require some modification of the CoinOR codes. It may be easier to work with this updated format for future integration.